Move global detail page component (#178)
