+++ frontend/app/routes/home.global.$volumeId/components/GlobalBookDetailAuthorBadge.tsx
