+++ frontend/app/routes/home.global.$volumeId/components/GlobalBookDetailContentTable.tsx
