+++ frontend/app/routes/home.global.$volumeId/components/GlobalBookDetailControlButtons.tsx
