+++ frontend/app/routes/home.global.$volumeId/components/GlobalBookDetailLink.tsx
