+++ frontend/app/routes/home.global.$volumeId/route.tsx
@@ -10,12 +10,12 @@ import { createBook, getBooks, getGoogleBook } from 'client/client';
 import { CreateBookBody, GoogleBook } from 'client/client.schemas';
 import { AiOutlineGlobal } from 'react-icons/ai';
 import { FaBookAtlas } from 'react-icons/fa6';
-import BookDetailActionPanel from '~/components/book-detail/BookDetailActionPanel';
 import BreadCrumbsComponent from '~/components/common/breadcrumbs/BreadCrumbsComponent';
-import GlobalBookDetailContent from '~/components/global-book-detail/GlobalBookDetailContent';
 import { commitSession, getSession } from '~/services/session.server';
 import { ActionResponse } from '~/types/response';
 import { makeCookieHeader } from '~/utils/session';
+import BookDetailActionPanel from './components/GlobalBookDetailActionPanel';
+import GlobalBookDetailContent from './components/GlobalBookDetailContent';
 
 interface LoaderData {
 	googleBook: GoogleBook;
